% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_continuous_data}
\alias{generate_continuous_data}
\alias{generate_binomial_data}
\alias{generate_count_data}
\title{Generate Data for Dose Response}
\usage{
generate_continuous_data(dose, fn, error_sd = 1, ...)

generate_binomial_data(dose, trials, fn, ...)

generate_count_data(dose, fn, ...)
}
\arguments{
\item{dose}{A numeric vector of doses. The number of observations is the length of this vector}

\item{fn}{A function for the mean response}

\item{error_sd}{Standard deviation of the error}

\item{...}{Additional arguments passed to \code{fn} function}

\item{trials}{The number of trials}
}
\value{
A data frame with two columns: \code{dose} and \code{response}, where \code{response} is the
calculated response for each dose, including the added error.
}
\description{
These functions generate continuous (normally-distributed), binomial, and count (Poisson) data.
}
\examples{
my_fn <- function(dose) { dose^2 } # example dose-response function
data <- generate_continuous_data(1:10, my_fn, error_sd = 0.5)
print(data)
}
