% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{log_logistic}
\alias{log_logistic}
\title{Log-Logistic Function}
\usage{
log_logistic(x, c, d, e, b)
}
\arguments{
\item{x}{A numeric vector of doses}

\item{c}{The lower limit}

\item{d}{The upper limit}

\item{e}{The EC/IC50, the response halfway between the upper and lower limit}

\item{b}{The Hill slope}
}
\value{
A numeric vector of mean responses
}
\description{
Main model used to generate mean responses. Most commonly used in fitting dose-respones relationships
}
