% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_drc_with_confidence}
\alias{plot_drc_with_confidence}
\title{Plot Results}
\usage{
plot_drc_with_confidence(drc_model, dose_column, response_column)
}
\arguments{
\item{response_column}{}
}
\description{
Plot Results
}
