% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dose_response_analysis}
\alias{dose_response_analysis}
\title{Dose Response Analysis}
\usage{
dose_response_analysis(
  dat,
  response,
  dose,
  type = c("stimulation", "inhibition"),
  vary_slope = FALSE,
  model_type = "log-logistic",
  log_transform_dose = FALSE,
  constant = 0.01,
  normalized_response = FALSE,
  dose_log = c("none", "log10", "natural"),
  family = c("continuous", "binomial"),
  fixed_params = NULL,
  constraints = list()
)
}
\arguments{
\item{constraints}{}
}
\description{
Dose Response Analysis
}
